+++ history.md
@@ -1,3 +1,4 @@
+* Fixed a bug in the gear loader where everything is True, even False things are True
 * Added Stream music mode
 * Fixed bug when lancemate with pet quits party
 * Fixed bug when reading black market sign with a pet in your party